% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_us_mortality_data.r
\name{read_us_mortality_data}
\alias{read_us_mortality_data}
\title{Read US mortality data}
\usage{
read_us_mortality_data(file_path)
}
\arguments{
\item{file_path}{the path to the file}
}
\value{
a tibble
}
\description{
Read US mortality data
}
\examples{
ii <- read_us_mortality_data(file_path)
}
